delete manage user assigned to project relation use case, add the functions on the manage user use case and handle its usage
