+++ src/main/kotlin/di/UseCaseModule.kt
@@ -9,7 +9,7 @@ import org.example.logic.usecase.audit.utils.AuditDescriptionProvider
 import org.example.logic.usecase.project.*
 import org.example.logic.usecase.state.ManageEntityStatesUseCase
 import org.example.logic.usecase.task.ManageTasksUseCase
-import org.example.logic.usecase.user.AddUserUseCase
+import org.example.logic.usecase.user.ManageUserUseCase
 import org.example.logic.utils.DateHandler
 import org.example.logic.utils.DateHandlerImp
 import org.example.logic.utils.HashingService
@@ -22,7 +22,7 @@ val useCaseModule = module {
     single<HashingService> { Md5HashingService() }
     single<DateHandler> { DateHandlerImp() }
 
-    single<AddUserUseCase> { AddUserUseCase(get(), get(), get(), get()) }
+    single<ManageUserUseCase> { ManageUserUseCase(get(), get(), get(), get(), get()) }
     single<GetAuditUseCase> { GetAuditUseCase(get(), get(), get()) }
     single<AddAuditUseCase> { AddAuditUseCase(get(), get()) }
     single<AuditServicesUseCase> { AuditServicesUseCase(get(), get(), get()) }
@@ -32,7 +32,6 @@ val useCaseModule = module {
     factory<ManageProjectUseCase> { ManageProjectUseCase(get(), get(), get()) }
     factory<ProjectValidationUseCase> { ProjectValidationUseCase(get()) }
     factory<GetProjectsUseCase> { GetProjectsUseCase(get()) }
-    factory<ManageUsersAssignedToProjectUseCase> { ManageUsersAssignedToProjectUseCase(get(), get(), get()) }
     factory<ManageEntityStatesUseCase> { ManageEntityStatesUseCase(get()) }
     factory<ValidateUserDataUseCase> { ValidateUserDataUseCase() }
 }
