+++ src/main/kotlin/logic/usecase/user/ManageUserUseCase.kt
@@ -4,16 +4,20 @@ import logic.usecase.login.LoginUseCase
 import logic.usecase.validation.ValidateUserDataUseCase
 import org.example.logic.InvalidPasswordException
 import org.example.logic.InvalidUserNameException
+import org.example.logic.UserDoesNotExistException
 import org.example.logic.UserExistException
 import org.example.logic.entities.User
 import org.example.logic.repository.UserRepository
+import org.example.logic.usecase.project.GetProjectsUseCase
 import org.example.logic.utils.HashingService
+import java.util.*
 
-class AddUserUseCase(
+class ManageUserUseCase(
     private val userRepository: UserRepository,
     private val validateUserDataUseCase: ValidateUserDataUseCase,
     private val loginUseCase: LoginUseCase,
     private val hashingService: HashingService,
+    private val getProjectsUseCase: GetProjectsUseCase,
 ) {
 
     suspend fun addUser(username: String, password: String): Boolean {
@@ -24,10 +28,31 @@ class AddUserUseCase(
         return userRepository.addUser(user)
     }
 
-    private suspend fun createUserIfNotExist(
-        username: String,
-        password: String,
-    ): User {
+    suspend fun addUserToProject(projectId: UUID, userName: String): Boolean {
+        return loginUseCase.isUserExist(userName).let { isUserExist ->
+            if (isUserExist)
+                userRepository.addUserToProject(projectId = projectId, username = userName)
+            else
+                throw UserDoesNotExistException()
+        }
+    }
+
+    suspend fun deleteUserFromProject(projectName: String, username: String): Boolean {
+        return getProjectsUseCase.getProjectByName(projectName).let { project ->
+            loginUseCase.isUserExist(username).let { isUserExist ->
+                if (isUserExist)
+                    userRepository.deleteUserFromProject(projectId = project.id, username = username)
+                else
+                    throw UserDoesNotExistException()
+            }
+        }
+    }
+
+    suspend fun getUsersByProjectId(projectId: UUID): List<User> {
+        return userRepository.getUsersByProjectId(projectId = projectId)
+    }
+
+    private suspend fun createUserIfNotExist(username: String, password: String): User {
         loginUseCase.isUserExist(username).also { exist ->
             if (!exist) {
                 return User(username = username, hashedPassword = hashingService.hash(password))
