+++ src/main/kotlin/ui/features/addusertoproject/AddUserToProjectUI.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.runBlocking
 import org.example.logic.entities.User
 import org.example.logic.usecase.project.GetProjectsUseCase
-import org.example.logic.usecase.project.ManageUsersAssignedToProjectUseCase
+import org.example.logic.usecase.user.ManageUserUseCase
 import org.example.ui.features.common.ui_launcher.UiLauncher
 import org.example.ui.features.common.utils.UiMessages
 import org.example.ui.features.user.CreateUserUi
@@ -15,7 +15,7 @@ class AddUserToProjectUI(
     private val inputReader: InputReader,
     private val outputPrinter: OutputPrinter,
     private val getProjectsUseCase: GetProjectsUseCase,
-    private val manageUsersAssignedToProjectUseCase: ManageUsersAssignedToProjectUseCase,
+    private val manageUserUseCase: ManageUserUseCase,
     private val createUserUiImp: CreateUserUi,
 ) : UiLauncher {
 
@@ -64,7 +64,7 @@ class AddUserToProjectUI(
         runBlocking(errorHandler) {
             try {
                 handleAssignUserToProjectMessage(
-                    manageUsersAssignedToProjectUseCase.addUserToProject(
+                    manageUserUseCase.addUserToProject(
                         getProjectsUseCase.getProjectByName(projectName).id, username
                     )
                 )
@@ -94,7 +94,7 @@ class AddUserToProjectUI(
             runBlocking(errorHandler) {
                 try {
                     handleUsersAssignedToProjectMessages(
-                        manageUsersAssignedToProjectUseCase.getUsersByProjectId(
+                        manageUserUseCase.getUsersByProjectId(
                             getProjectsUseCase.getProjectByName(projectName).id
                         ),
                         projectName
@@ -125,7 +125,7 @@ class AddUserToProjectUI(
                 outputPrinter.showMessage("${UiMessages.ENTER_USER_NAME_TO_REMOVE_PROJECT} ${UiMessages.OR_LEAVE_IT_BLANK_TO_BACK}")
                 inputReader.readStringOrNull()?.let { username ->
                     try {
-                        manageUsersAssignedToProjectUseCase.deleteUserFromProject(projectName, username)
+                        manageUserUseCase.deleteUserFromProject(projectName, username)
                         outputPrinter.showMessageLine(UiMessages.USER_DELETED_FROM_PROJECT)
                     } catch (e: Exception) {
                         outputPrinter.showMessageLine("${UiMessages.FAILED_TO_DELETE_USER_FROM_PROJECT} ${e.message}")
