+++ src/test/kotlin/ui/features/project/ProjectManagerUiImpTest.kt
@@ -1,8 +1,10 @@
 package ui.features.project
 
 import org.example.logic.usecase.project.ManageProjectUseCase
-import org.example.logic.usecase.project.ManageUsersAssignedToProjectUseCase
+import org.example.logic.usecase.user.ManageUserUseCase
 import org.example.ui.features.project.ProjectManagerUi
+import org.example.ui.features.state.AdminEntityStateManagerUi
+import org.example.ui.features.task.TaskManagerUi
 import org.example.ui.features.user.CreateUserUi
 import org.example.ui.input_output.input.InputReader
 import org.example.ui.input_output.output.OutputPrinter
@@ -15,7 +17,7 @@ class ProjectManagerUiImpTest {
     private lateinit var inputReader: InputReader
     private lateinit var outputPrinter: OutputPrinter
     private lateinit var projectManagerUi: ProjectManagerUi
-    private lateinit var manageUsersAssignedToProjectUseCase: ManageUsersAssignedToProjectUseCase
+    private lateinit var manageUserUseCase: ManageUserUseCase
 
 
 //    @BeforeEach
