+++ src/test/kotlin/ui/features/user/CreateUserUiTest.kt
@@ -1,29 +1,27 @@
 package ui.features.user
 
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.verify
-import logic.model.entities.User
-import logic.model.entities.UserRole
-import org.example.logic.model.exceptions.UsersDataAreEmpty
-import org.example.logic.usecase.extention.toSafeUUID
+import io.mockk.*
+import kotlinx.coroutines.test.runTest
+import org.example.logic.UsersDataAreEmptyException
+import org.example.logic.entities.User
+import org.example.logic.entities.UserRole
+import org.example.logic.usecase.user.ManageUserUseCase
+import org.example.ui.features.user.CreateUserUi
 import org.example.ui.input_output.input.InputReader
 import org.example.ui.input_output.output.OutputPrinter
-import org.example.logic.usecase.user.AddUserUseCase
-import org.example.ui.features.user.CreateUserUi
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.util.*
 
 class CreateUserUiTest {
-    private val mockCreateUserUseCase: AddUserUseCase = mockk(relaxed = true)
+    private val mockCreateUserUseCase: ManageUserUseCase = mockk(relaxed = true)
     private val mockPrinter: OutputPrinter = mockk(relaxed = true)
     private val mockInputReader: InputReader = mockk()
     private lateinit var createUserUi: CreateUserUi
     val user = User(
         username = "rodina",
         hashedPassword = "0192023a7bbd73250516f069df18b500",
-        id = "de25ea95-fca4-4f87-90e8-fd9c6aa0a716".toSafeUUID(),
+        id = UUID.fromString("de25ea95-fca4-4f87-90e8-fd9c6aa0a716"),
         userRole = UserRole.ADMIN
     )
 
@@ -39,7 +37,7 @@ class CreateUserUiTest {
         every { mockInputReader.readStringOrNull() } returnsMany listOf(null, null)
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
         verify {
@@ -55,7 +53,7 @@ class CreateUserUiTest {
         every { mockInputReader.readStringOrNull() } returnsMany listOf("", "password")
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
         verify { mockPrinter.showMessageLine("Username and password cannot be empty") }
@@ -67,7 +65,7 @@ class CreateUserUiTest {
         every { mockInputReader.readStringOrNull() } returnsMany listOf("username", "")
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
         verify { mockPrinter.showMessageLine("Username and password cannot be empty") }
@@ -79,7 +77,7 @@ class CreateUserUiTest {
         every { mockInputReader.readStringOrNull() } returnsMany listOf(null, null)
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
         verify { mockPrinter.showMessageLine("Username and password cannot be empty") }
@@ -87,93 +85,90 @@ class CreateUserUiTest {
 
 
     @Test
-    fun `should show success message when user added successfully`() {
+    fun `should show success message when user added successfully`() = runTest {
         // Given
         val username = "testUser"
         val password = "testPass123"
         every { mockInputReader.readStringOrNull() } returnsMany listOf(username, password)
-        every { mockCreateUserUseCase.addUser(username, password) } returns Result.success(true)
+        coEvery { mockCreateUserUseCase.addUser(username, password) } returns true
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
         verify { mockPrinter.showMessageLine("✅ User $username added successfully!") }
     }
 
     @Test
-    fun `should show error message when user addition fails`() {
+    fun `should show error message when user addition fails`() = runTest {
         // Given
         val username = "testUser"
         val password = "testPass123"
         every { mockInputReader.readStringOrNull() } returnsMany listOf(username, password)
-        every { mockCreateUserUseCase.addUser(username, password) } returns Result.success(false)
+        coEvery { mockCreateUserUseCase.addUser(username, password) } returns false
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
         verify { mockPrinter.showMessageLine("Failed to add user") }
     }
 
     @Test
-    fun `should show error message when exception occurs`() {
+    fun `should show error message when exception occurs`() = runTest {
         // Given
         val username = "testUser"
         val password = "testPass123"
         every { mockInputReader.readStringOrNull() } returnsMany listOf(username, password)
-        every { mockCreateUserUseCase.addUser(username, password) } returns
-                Result.failure(UsersDataAreEmpty())
+        coEvery { mockCreateUserUseCase.addUser(username, password) } throws UsersDataAreEmptyException()
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
-        verify { mockPrinter.showMessageLine("${UsersDataAreEmpty().message}") }
+        verify { mockPrinter.showMessageLine("${UsersDataAreEmptyException().message}") }
     }
 
 
     @Test
-    fun `should call addUserUseCase with correct parameters`() {
+    fun `should call addUserUseCase with correct parameters`() = runTest {
         // Given
         val username = "testUser"
         val password = "testPass123"
         every { mockInputReader.readStringOrNull() } returnsMany listOf(username, password)
-        every { mockCreateUserUseCase.addUser(username, password) } returns Result.success(true)
+        coEvery { mockCreateUserUseCase.addUser(username, password) } returns true
 
         // When
-        createUserUi.launchUi(
-            user
-        )
+        createUserUi.launchUi()
 
 
         // Then
-        verify { mockCreateUserUseCase.addUser(username, password) }
+        coVerify { mockCreateUserUseCase.addUser(username, password) }
     }
 
     @Test
-    fun `should reject null username`() {
+    fun `should reject null username`() = runTest {
         // Given
         every { mockInputReader.readStringOrNull() } returnsMany listOf(null, "validPass")
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
-        verify(exactly = 0) { mockCreateUserUseCase.addUser(any(), any()) }
+        coVerify(exactly = 0) { mockCreateUserUseCase.addUser(any(), any()) }
         verify { mockPrinter.showMessageLine("Username and password cannot be empty") }
     }
 
     @Test
-    fun `should reject null password`() {
+    fun `should reject null password`() = runTest {
         // Given
         every { mockInputReader.readStringOrNull() } returnsMany listOf("validUser", null)
 
         // When
-        createUserUi.launchUi(user)
+        createUserUi.launchUi()
 
         // Then
-        verify(exactly = 0) { mockCreateUserUseCase.addUser(any(), any()) }
+        coVerify(exactly = 0) { mockCreateUserUseCase.addUser(any(), any()) }
         verify { mockPrinter.showMessageLine("Username and password cannot be empty") }
     }
